backend/requirements.txt
django>=4.2
djangorestframework
django-cors-headers
google-generativeai
psycopg2-binary
python-decouple
